Version 1.0 functionality:  Fixing deleting MicroLessons.

+++ qpalx-elearning-web-zone-content-admin/src/main/java/com/quaza/solutions/qpalx/elearning/web/zone/content/admin/curriculum/QPalXLessonAdminController.java
@@ -321,7 +321,9 @@ public class QPalXLessonAdminController {
             iRedirectStrategyExecutor.sendRedirectWithError(targetURL, error, WebOperationErrorAttributesE.Invalid_Delete_Operation, request, response);
         } else {
             iqPalXELessonService.deleteQPalXELesson(qPalXELesson);
-            ieLearningStaticContentService.deleteELearningMediaContent(qPalXELesson.geteLearningMediaContent());
+            if (qPalXELesson.geteLearningMediaContent() != null) {
+                ieLearningStaticContentService.deleteELearningMediaContent(qPalXELesson.geteLearningMediaContent());
+            }
             String targetURL = "/view-admin-qpalx-elessons?eLearningCourseID=" + eLearningCourse.getId();
             iRedirectStrategyExecutor.sendRedirect(request, response, targetURL);
         }
